backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "outreach-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Auth
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",

    # Config
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.10.0",

    # Task queue
    "arq>=0.26.1",
    "redis>=5.2.0",

    # AI
    "anthropic>=0.40.0",

    # Email / OAuth
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.154.0",
    "msal>=1.31.0",
    "httpx>=0.28.0",

    # DNS verification
    "dnspython>=2.7.0",

    # CSV processing
    "python-csv>=0.0.1",
    "chardet>=5.2.0",

    # Utilities
    "cryptography>=44.0.0",
    "python-dateutil>=2.9.0",
    "orjson>=3.10.0",
    "playwright>=1.49.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-respx>=0.22.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.1",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-passlib",
    "types-python-jose",
    "types-python-dateutil",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
omit = ["tests/*", "alembic/*"]
